use SNAP_USER_COMMON to store .librewolf
This directory is for permanent data. It does not get backed up between releases, so some storage will be saved with this change.

However, in case of a bad update, this makes reverting more difficult because firefox does not let you use a profile from version x+1 on version x.